refactor: extract AgentTypeRegistry to models layer
Move agent-type registry out of the service into a simple class
instance at models/agent-type-registry/. The service consumes it
via agentTypeRegistry.get(agentType).
