+++ apps/backend/src/models/agent-type-registry/index.ts
@@ -1,0 +1,5 @@
+export {
+  type AgentConstructor,
+  type AgentSessionStore,
+  agentTypeRegistry,
+} from './agent-type-registry.js';
