+++ apps/backend/src/services/agent-session/index.ts
@@ -1,10 +1,3 @@
-export {
-  agentSessionService,
-  registerAgentType,
-} from './agent-session-service.js';
-export type {
-  AgentConstructor,
-  AgentSessionStore,
-  CreateSessionResult,
-} from './types.js';
+export {agentSessionService} from './agent-session-service.js';
+export type {CreateSessionResult} from './types.js';
 export {CreateSessionError} from './types.js';
